chore: ignore some Snyk issues
Ignore a new issue in a transitive dependency, extend the expiry dates
of 2 other ignore rules, and ignore 4 more that had been issues for some
time. All of these packages are transitive dependencies of docusaurus,
which is only used to generate a static site, and does not take user
input, so we think none of the issues are reachable.